whatsapp-chatbot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.1.0
langchain-groq==0.0.1
langchain-community==0.0.10
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
python-dateutil==2.8.2 